+++ lib/assertion/guard.rb
@@ -34,12 +34,6 @@ module Assertion
 
     extend GuardDSL
 
-    # @!attribute [r] object
-    #
-    # @return [Object] The object whose state should be tested
-    #
-    attr_reader :object
-
     # @!scope class
     # @!method new(object)
     # Creates the guard instance for the provided object